+++ CHANGELOG.md
@@ -13,6 +13,10 @@ It is expected that you keep this format strictly, since we depend on it in our 
 
 - `SBBTextField`: Remove listener on (custom) FocusNode and dispose own TextEditingController (#392) (#396)
 
+### Changed
+
+- (auto): updated icon lib to version 1.8.10
+
 ## [4.2.0] - 2025-10-03
 
 ### Added