+++ CHANGELOG.md
@@ -15,7 +15,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 
 ### Changed
 
-- (auto): updated icon lib to version 1.8.10
+- (auto): updated icon lib to version 1.8.11
 
 ## [4.2.0] - 2025-10-03