+++ CHANGELOG.md
@@ -12,6 +12,9 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Fixed
 
 - `SBBTextField`: Remove listener on (custom) FocusNode and dispose own TextEditingController (#392) (#396)
+- `SBBSegmentedButton`: Fixed animation issue when resizing SBBSegmentedButton (#312)
+- `SBBSegmentedButton`: Removed `boxShadow` from `SBBSegmentedButtonStyle` (#312)
+- `SBBSegmentedButton`: Moved `iconColor` from `SBBSegmentedButtonStyle` to `SBBSegmentedButtonInnerStyle` (#312)
 
 ### Changed