+++ CHANGELOG.md
@@ -12,6 +12,13 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Fixed
 
 - `SBBTextField`: Remove listener on (custom) FocusNode and dispose own TextEditingController (#392) (#396)
+- `SBBSegmentedButton`: Fixed animation issue when resizing SBBSegmentedButton (#312)
+- `SBBSegmentedButton`: Removed `boxShadow` from `SBBSegmentedButtonStyle` (#312)
+- `SBBSegmentedButton`: Moved `iconColor` from `SBBSegmentedButtonStyle` to `SBBSegmentedButtonInnerStyle` (#312)
+
+### Changed
+
+- (auto): updated icon lib to version 1.8.11
 
 ## [4.2.0] - 2025-10-03