+++ CHANGELOG.md
@@ -24,6 +24,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Changed
 
 - (auto): updated icon lib to version 1.8.11
+- BREAKING: removed default left margin on component `SBBTextField` (#402)
 
 ## [4.2.0] - 2025-10-03